Create test-helper
Extract beforeEach to test-helper
Remove afterEach, as before each rollsback migration. beforeEach has to rollback the migration because if the test fails it won't reach the afterEach to reset the database. This might be less performant on larger applications.
Delete xml-schema, obsolete.
Add notes to removed and replaced with project management tool.